compilebench: add torture tests
Add a framework for benchmarking compiler torture tests.
Add two recent torture tests. There are others to harvest,
but this is a start.

Having to unzip the files every time is unfortunate,
but (a) the compilation is crazy slow anyway,
(b) the torture files tend to be large, and
(c) the compression factor is ~10x.